+++ src/core/univ3forks/SwapModeV3.sol
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.25;
+
+address constant swapModeV3Factory = 0x6E36FC34eA123044F278d3a9F3819027B21c9c32;
+bytes32 constant swapModeV3InitHash = 0x4adf6002ac7c691132e09e25068a1ddc77accbf11c16b012cc95b7d1842a0af9;
+uint8 constant swapModeV3ForkId = 23;